Moved get_ssl_ methods to base class and fixes for logger and tls interface
